fix: enable API docs versioning for Python projects (#47)
Moves Python logic directly to the `generateJson` method, so it gets
called even when `docusaurus api:version` is called.

Expands the `Unpack[]` - annotated parameters in the JSON-generation
stage (before, this was done during the rendering phase), so that
reexported symbols have their `Unpack`ed arguments.

This also helps with rendering the `Unpack`ed kwargs in the rendered
method signatures.